added receipt point calculator and working on test

+++ src/main/java/org/example/models/ReceiptsPoints.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReceiptsPoints {
-    public Integer points;
+    public Double points;
 }
